backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
bcrypt
python-jose
tensorflow
Pillow
cryptography
pytz